requirements.txt
absl-py==1.3.0
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
customtkinter==5.0.3
cycler==0.11.0
darkdetect==0.8.0
flatbuffers==22.12.6
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
mediapipe==0.9.0.1
numpy==1.24.1
opencv-contrib-python==4.7.0.68
packaging==22.0
Pillow==9.3.0
protobuf==3.20.3
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
